requirements.txt
asgiref==3.5.2
Django==4.0.6
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-formtools==2.3
gunicorn==20.1.0
numpy==1.23.1
pandas==1.4.3
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
